test: extract SolverContractTests abstract base
